Add_PaymentCard

+++ Project1/Project1/MovePositionCard.cpp
@@ -1,6 +1,6 @@
 #include "MovePositionCard.h"
 
-MovePositionCard::MovePositionCard(const MyString& name, const MyString& effect, int moveWith)	
+MovePositionCard::MovePositionCard(const MyString& name, const MyString& effect, size_t moveWith)
 	:Card(name, effect)
 {
 	this->moveWith = moveWith;
